test(workflow): restore blocked specs and switch semver utils to runtime-safe package
